refactor: extract Oid alias + PG constants to crate-root module
Before, src/types/sql_type.rs imported `crate::catalog::oid::Oid` — a
reverse dependency, since catalog is conceptually a lower layer and
types carries wire-protocol semantics on top. Mirror side: catalog
holds a `types::SqlType` inside Column, which is the up-direction
dependency. The Oid alias move breaks the types→catalog direction
without touching the Column.sql_type decision, which is a separate
design choice.

Changes:
- new `src/oid.rs` holds `Oid`, `FIRST_NORMAL_OID`, the canonical PG
  namespace/superuser OIDs. These are primitive identifiers, not
  catalog state, so a leaf module is the right home.
- `src/catalog/oid.rs` re-exports those and keeps `OidGenerator`
  (catalog-specific state with `CatalogError`) unchanged.
- `src/types/sql_type.rs` switches to `crate::oid::Oid`.

All 12 other files that import `catalog::oid::{…}` keep working
unchanged via the re-export. The Column.sql_type inversion is the
larger architectural call and is tracked separately.
