+++ src/catalog/oid.rs
@@ -1,18 +1,12 @@
 use super::CatalogError;
 
-pub type Oid = u32;
-
-pub const FIRST_NORMAL_OID: Oid = 16_384;
-
-// PG-canonical namespace OIDs. These match Postgres's bootstrap constants so
-// that catalog JOINs like `pg_type.typnamespace = pg_namespace.oid` hit the
-// rows real drivers expect.
-pub const PG_CATALOG_NAMESPACE_OID: Oid = 11;
-pub const PUBLIC_NAMESPACE_OID: Oid = 2_200;
-pub const INFORMATION_SCHEMA_NAMESPACE_OID: Oid = 13_000;
-
-// Superuser-role OID used wherever pg_catalog rows need an owner.
-pub const BOOTSTRAP_SUPERUSER_OID: Oid = 10;
+// Re-exports from the crate-root `oid` module. Kept here so existing
+// callers that import `crate::catalog::oid::Oid` continue to work after
+// the leaf module was introduced (Phase 2 layering cleanup).
+pub use crate::oid::{
+    BOOTSTRAP_SUPERUSER_OID, FIRST_NORMAL_OID, INFORMATION_SCHEMA_NAMESPACE_OID, Oid,
+    PG_CATALOG_NAMESPACE_OID, PUBLIC_NAMESPACE_OID,
+};
 
 #[derive(Debug, Clone)]
 pub struct OidGenerator {
