+++ src/lib.rs
@@ -67,6 +67,7 @@ pub mod commands;
 pub mod executor;
 pub mod extensions;
 pub mod foreign;
+pub mod oid;
 pub mod parser;
 pub mod planner;
 pub mod plpgsql;
